requirements.txt
yfinance>=0.2.0
pandas>=1.0.0
requests>=2.25.0
PyYAML>=5.0
beautifulsoup4>=4.9.0
cloudscraper>=1.2.0
